+++ python/marshmallow_recipe/fields.py
@@ -13,7 +13,7 @@ from typing import Any
 import marshmallow as m
 import marshmallow.validate
 
-from .validation import ValidationFunc, combine_validators
+from .validation import ValidationFunc, combine_validators, regexp_validate
 
 _MARSHMALLOW_VERSION_MAJOR = int(importlib.metadata.version("marshmallow").split(".")[0])
 
@@ -28,6 +28,8 @@ def str_field(
     min_length_error: str | None = None,
     max_length: int | None = None,
     max_length_error: str | None = None,
+    regexp: str | None = None,
+    regexp_error: str | None = None,
     validate: ValidationFunc | collections.abc.Sequence[ValidationFunc] | None = None,
     strip_whitespaces: bool = False,
     post_load: collections.abc.Callable[[str], str] | None = None,
@@ -41,6 +43,9 @@ def str_field(
         validate = combine_validators(validate, m.validate.Length(min=min_length, error=min_length_error))
     if max_length is not None:
         validate = combine_validators(validate, m.validate.Length(max=max_length, error=max_length_error))
+    if regexp is not None:
+        error = (regexp_error or "String does not match expected pattern.").replace("{", "{{").replace("}", "}}")
+        validate = combine_validators(validate, regexp_validate(regexp, error=error))
     if default is m.missing:
         return StrField(
             allow_none=allow_none,